+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `hideUnavailableItems` prop, along with accompanying option on the Shelf schema.
 
 ## [1.21.2] - 2019-07-25
 ### Changed